# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Redstone Squid API",
    "description": "Versioned public API for the Redstone Squid build catalog.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/users/me/consent": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Grant Consent",
        "description": "Accept the current privacy notice for future writes.",
        "operationId": "grant_consent_v1_users_me_consent_post",
        "security": [
          {
            "ApiCredential": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "pattern": "^[\\x21-\\x7e]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours.",
              "title": "Idempotency-Key"
            },
            "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/vote-sessions/{vote_session_id}": {
      "get": {
        "tags": [
          "vote sessions"
        ],
        "summary": "Get Vote Session",
        "description": "Return aggregate vote state without exposing ballot identities.",
        "operationId": "get_vote_session_v1_vote_sessions__vote_session_id__get",
        "security": [
          {
            "ApiCredential": []
          }
        ],
        "parameters": [
          {
            "name": "vote_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vote Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteSessionDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/v1/vote-sessions/{vote_session_id}/votes": {
      "post": {
        "tags": [
          "vote sessions"
        ],
        "summary": "Cast Vote",
        "description": "Cast the authenticated Discord member's weighted vote.",
        "operationId": "cast_vote_v1_vote_sessions__vote_session_id__votes_post",
        "security": [
          {
            "ApiCredential": []
          }
        ],
        "parameters": [
          {
            "name": "vote_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Vote Session Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255,
                  "pattern": "^[\\x21-\\x7e]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours.",
              "title": "Idempotency-Key"
            },
            "description": "Deduplicate an equivalent mutation by this authenticated caller for 24 hours."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteSessionDetail"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "ACCOUNT_ALREADY_LINKED",
          "ALIAS_ALREADY_ADDED",
          "ALIAS_ALREADY_CLAIMED",
          "ALIAS_IN_USE",
          "BUILD_BUSY",
          "BUILD_NOT_FOUND",
          "BUILD_REVISION_MISMATCH",
          "BUILD_REVISION_REQUIRED",
          "CLAIM_NOT_FOUND",
          "CREATOR_ALIAS_NOT_FOUND",
          "CONFIGURATION_ERROR",
          "CONSENT_REQUIRED",
          "DATA_INTEGRITY_ERROR",
          "DOMAIN_ERROR",
          "INFRASTRUCTURE_ERROR",
          "IDEMPOTENCY_CONFLICT",
          "IDEMPOTENCY_IN_PROGRESS",
          "INTERNAL_ERROR",
          "INVALID_BUILD",
          "INVALID_CURSOR",
          "INVALID_MESSAGE",
          "INVALID_QUERY",
          "INVALID_REQUEST",
          "INVALID_STATE",
          "INVALID_USER",
          "INVALID_VERIFICATION_CODE",
          "INVALID_VERSION",
          "INVALID_VOTE_CONFIGURATION",
          "MESSAGE_NOT_FOUND",
          "MINECRAFT_ACCOUNT_NOT_FOUND",
          "MINECRAFT_SERVICE_UNAVAILABLE",
          "NOT_FOUND",
          "PERSISTENCE_ERROR",
          "RATE_LIMITED",
          "RESTRICTION_NOT_FOUND",
          "SCHEMATIC_INVALID",
          "SCHEMATIC_NOT_FOUND",
          "SCHEMATIC_RENDER_UNAVAILABLE",
          "SCHEMATIC_SUPPORT_UNAVAILABLE",
          "SCHEMATIC_TIMEOUT",
          "SCHEMATIC_TOO_LARGE",
          "SCHEMATIC_WORKER_CRASHED",
          "UNAUTHORIZED",
          "USER_NOT_FOUND",
          "VALIDATION_ERROR",
          "VERSION_CATALOG_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "description": "Stable machine-readable application error codes."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JSONValue": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JSONValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "OwnVoteSelection": {
        "properties": {
          "option_id": {
            "type": "string",
            "title": "Option Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "option_id"
        ],
        "title": "OwnVoteSelection",
        "description": "The authenticated caller's own ballot selection."
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JSONValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "error_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "description": "RFC 9457 problem detail response with application extensions."
      },
...
      "UserMe": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "discord_id": {
            "type": "integer",
            "title": "Discord Id"
          },
          "minecraft_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minecraft Uuid"
          },
          "ign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ign"
          },
          "consent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Version"
          },
          "consent_pending": {
            "type": "boolean",
            "title": "Consent Pending"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "discord_id",
          "minecraft_uuid",
          "ign",
          "consent_version",
          "consent_pending"
        ],
        "title": "UserMe",
        "description": "The caller's own linked-account data."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VoteInput": {
        "properties": {
          "guild_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Guild Id"
          },
          "option_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Option Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "guild_id",
          "option_id"
        ],
        "title": "VoteInput",
        "description": "A stable option selection in one Discord guild."
      },
      "VoteOptionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "choice": {
            "type": "string",
            "title": "Choice"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "choice",
          "position"
        ],
        "title": "VoteOptionSummary",
        "description": "A stable vote option without Discord-specific reaction aliases."
      },
      "VotePollSummary": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "visibility",
          "deadline"
        ],
        "title": "VotePollSummary",
        "description": "Public generic-poll metadata."
      },
      "VoteSessionDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "type": "string",
            "title": "Result"
          },
          "pass_threshold": {
            "type": "integer",
            "title": "Pass Threshold"
          },
          "fail_threshold": {
            "type": "integer",
            "title": "Fail Threshold"
          },
          "build_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Id"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/VoteOptionSummary"
            },
            "type": "array",
            "title": "Options"
          },
          "tallies": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoteTallies"
              },
              {
                "type": "null"
              }
            ]
          },
          "poll": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VotePollSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "own_selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnVoteSelection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "result",
          "pass_threshold",
          "fail_threshold",
          "build_id",
          "options",
          "tallies",
          "poll",
          "own_selection"
        ],
        "title": "VoteSessionDetail",
        "description": "A vote session with aggregate-only ballot data."
      },
      "VoteTallies": {
        "properties": {
          "raw": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Raw"
          },
          "weighted": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weighted"
          },
          "upvotes": {
            "type": "number",
            "title": "Upvotes"
          },
          "downvotes": {
            "type": "number",
            "title": "Downvotes"
          },
          "net": {
            "type": "number",
            "title": "Net"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "raw",
          "weighted",
          "upvotes",
          "downvotes",
          "net"
        ],
        "title": "VoteTallies",
        "description": "Aggregate results which cannot identify individual voters."
      }
    },
    "securitySchemes": {
      "ApiCredential": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
...
}